feat: Add secondary field support for TestHistoryCard
Enhances the `TestHistoryCard` component to support an optional secondary field. This allows for storing additional related information, such as the type of test (e.g., HCV Type).

Updates the `HbvHcvTab` to leverage this new functionality for storing HCV test types. Also includes a minor version bump in the LoginPage footer.